[tests] Use xtf file instead of database backup and update some tests in test_quality_rules

+++ asistente_ladm_col/gui/qgis_model_baker/dlg_import_schema.py
@@ -311,7 +311,7 @@ class DialogImportSchema(QDialog, DIALOG_UI):
         self.accepted()
 
     def download_java_progress_change(self, progress):
-        self.progress_bar.setValue(progress/2)
+        self.progress_bar.setValue(int(progress/2))
         if (progress % 20) == 0:
             self.txtStdout.append('...')
